update production with latest features

+++ src/AppBundle/Controller/ManageUsersController.php
@@ -8,10 +8,18 @@
 
 namespace AppBundle\Controller;
 
+use AppBundle\AppBundle;
+use AppBundle\Entity\User;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 
 
+/**
+ * Class ManageUsersController
+ * @package AppBundle\Controller
+ * @Security("is_granted('ROLE_ADMIN')")
+ */
 class ManageUsersController extends Controller
 {
 
@@ -25,6 +33,19 @@ class ManageUsersController extends Controller
 
         return $this->render("user/listUsers.html.twig", [
             'users' => $users
+        ]);
+    }
+
+    /**
+     * @Route("/list_users/{id}", name="list_posts_by_user")
+     */
+    public function listUsersPosts(User $user){
+        $em = $this->getDoctrine()->getManager();
+
+        $itemPosts = $em->getRepository("AppBundle:ItemPost")->findAllItemsByUser($user);
+
+        return $this->render("user/viewUsersPosts.html.twig", [
+            'itemPosts' => $itemPosts
         ]);
     }
 
@@ -50,10 +71,12 @@ class ManageUsersController extends Controller
         $em = $this->getDoctrine()->getManager();
         $user = $em->getRepository('AppBundle:User')->find($id);
 
-        $user->removeOneRole('ROLE_ADMIN');
+        if(!in_array('ROLE_SUPER_ADMIN', $user->getRoles())){
+            $user->removeOneRole('ROLE_ADMIN');
 
-        $em->persist($user);
-        $em->flush();
+            $em->persist($user);
+            $em->flush();
+        }
 
         return $this->redirectToRoute('list_users');
     }
@@ -65,8 +88,10 @@ class ManageUsersController extends Controller
         $em = $this->getDoctrine()->getManager();
         $user = $em->getRepository('AppBundle:User')->find($id);
 
-        $em->remove($user);
-        $em->flush();
+        if(!in_array('ROLE_SUPER_ADMIN', $user->getRoles())){
+            $em->remove($user);
+            $em->flush();
+        }
 
         return $this->redirectToRoute('list_users');
     }
